requirements.txt
opencv-python
pyzmq
numpy
pyaudio
